requirements.txt
python-pptx>=1.0.0
pyinstaller>=6.0.0
